Rename to config the MantraData instantiation in MantraForm class
Previously, MantraData was the name of the constant that store the object returned by the instantiation of the MantraData inside the MantraForm class. But there was a naming conflict because two parts of the MantraForm file declares the same name: 1. Import statement (which is a global constant) and 2. getConfig method (which is a local constant). The output result was an error because there can't be a global and local constants with the same name.

Currently, config is the name of the constant inside the getConfig method which don't produces the naming conflict inside the MantraForm file.