Set the component property of MantraData class
The component property of a MantraData class is set using static methods from the _baseClass property. The _baseClass is a component class that contains static methods to retrieve a component that the MantraData needs to set.

To achieve the set of the component property, the following changes were made:

In the MantraForm class:

- Added new imports of get and cloneDeep methods of lodash.

- Created a MIXIN constant to verify if a MantraComponent object contains the valid MantraForm mixin.

- Created the getComponent static method to retrieve a MantraForm component from a schema.

- Created the getPath static method to generate a path to traverse a schema.

- In the getConfig method, added a try/catch statement to handle the errors that may be produce by any of the MantraData's set methods.

In the MantraData class:

- Added the setComponent method that set the component property of the MantraData class.

- In the setBase method, added a throw error in case. no schema was found.

In the MantraComponent class:

- Removed the isValidComponent method from the class and let each MantraComponent implement it as a static method.

In the package json:

- Added lodash as a dependency to be used by the MantraForm class.