Removed the DATA_TYPES constant from the Factory module to a separate module
Previously, the Factory module had a constant reflecting all of the supported Mantra data type classes but this could cause circular dependencies if any of data type classes dependend on the Factory module. Currently, the DATA_TYPES constant is in the ./index module of the DataTypes folder and exports as a default a function that returns all the supported data types because functions don't produce circular dependencies.