Change the baseClass constructor param to be a string
Previously, the MantraData module was passed through a param the definition of the baseClass it will need to store as a property.  Currently, instead of a class definition, a string with the name of the class is passed as the baseClass parameter. This string will be used to search for the class definition and set it in the _baseClass property of MantraData. To avoid circular dependencies, the MantraData imports the getComponentClasses function to retrieve all supported MantraComponent classes in an object that we call BASE_CLASSES and from that object we select the baseClass the MantraData needs to store.